TravelApp/requirements.txt
app==0.0.1
click==6.6
Flask==0.11.1
geopy==1.11.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycountry==1.20
pysolr==3.5.0
python-dateutil==2.5.3
redis==2.10.5
requests==2.11.1
six==1.10.0
watson-developer-cloud==0.21.0
Werkzeug==0.11.11